requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# IBM watsonx Orchestrate ADK
ibm-watsonx-ai==0.2.0
ibm-watson==8.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.11  # PostgreSQL (Python 3.13 compatible)
aiosqlite==0.19.0  # SQLite
pymysql==1.1.0  # MySQL

# ML/AI Models
prophet==1.1.5
shap==0.44.1
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
polars==0.20.3

# API Clients
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1
google-search-results==2.4.2  # SerpAPI
praw==7.7.1  # Reddit API
google-api-python-client==2.116.0  # YouTube API

# Async & Task Queue
celery==5.3.6
redis==5.0.1
asyncio==3.4.3

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing async endpoints

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.3
faker==22.6.0  # For generating synthetic data
tqdm==4.66.1

# CORS
fastapi-cors==0.0.6

# Visualization (for reports)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
